requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=py38h32f6830_1
affine=2.3.0=py_0
astroid=2.4.2=py38_0
attrs=19.3.0=py_0
blas=1.0=mkl
blosc=1.20.0=he1b5a44_0
boost-cpp=1.72.0=h7b93d67_2
brotli=1.0.7=he1b5a44_1004
brotlipy=0.7.0=py38h7b6447c_1000
bzip2=1.0.8=h516909a_3
c-ares=1.16.1=h516909a_0
ca-certificates=2020.6.20=hecda079_0
cairo=1.16.0=h3fc0475_1005
cartopy=0.18.0=py38h4f3c6b6_2
certifi=2020.6.20=py38h924ce5b_2
cffi=1.14.1=py38he30daa8_0
cfitsio=3.470=hce51eda_6
chardet=3.0.4=py38_1003
charls=2.1.0=he1b5a44_2
click=7.1.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.5.0=py_0
cloudpickle=1.6.0=py_0
comet-git-pure=0.19.16=py_0
comet_ml=3.1.16=py38
configobj=5.0.6=py38_1
cryptography=2.9.2=py38h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
curl=7.71.1=he644dc0_5
cycler=0.10.0=py38_0
cytoolz=0.10.1=py38h516909a_0
dask-core=2.24.0=py_0
dbus=1.13.16=hb2f20db_0
decorator=4.4.2=py_0
everett=1.0.2=py_1
expat=2.2.9=he6710b0_2
fontconfig=2.13.1=h1056068_1002
freetype=2.10.2=h5ab3b9f_0
freexl=1.0.5=h516909a_1002
future=0.18.2=py38h32f6830_1
gdal=3.1.2=py38hb61cb63_1
geos=3.8.1=he1b5a44_0
geotiff=1.6.0=ha04d9d0_1
giflib=5.2.1=h516909a_2
glib=2.65.0=h3eb4bd4_0
grpcio=1.31.0=py38h2c89da0_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
hdf4=4.2.13=hf30be14_1003
hdf5=1.10.6=nompi_h3c11f04_101
icu=67.1=he1b5a44_0
idna=2.10=py_0
imagecodecs=2020.5.30=py38h36e1e94_1
imageio=2.9.0=py_0
importlib-metadata=1.7.0=py38h32f6830_0
importlib_metadata=1.7.0=0
intel-openmp=2020.1=217
isort=4.3.21=py38_0
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
json-c=0.13.1=hbfbb72e_1002
jsonschema=3.2.0=py38_0
jxrlib=1.1=h516909a_2
kealib=1.4.13=h33137a7_1
kiwisolver=1.2.0=py38hfd86e86_0
krb5=1.17.1=hfafb76e_2
lazy-object-proxy=1.4.3=py38h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libaec=1.0.4=he1b5a44_1
libcurl=7.71.1=hcdd3856_5
libdap4=3.20.6=h1d1bd15_1
libedit=3.1.20191231=h14c3975_1
libev=4.33=h516909a_0
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgdal=3.1.2=hb2a6f5f_1
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.16=h516909a_0
libkml=1.3.0=hb574062_1011
libllvm9=9.0.1=h4a3c616_1
libnetcdf=4.7.4=nompi_h84807e1_105
libnghttp2=1.41.0=h8cfc5f6_2
libpng=1.6.37=hbc83047_0
libpq=12.3=h5513abc_0
libprotobuf=3.12.4=h8b12597_0
libspatialite=4.3.0a=h57f1b35_1039
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=h68273f3_2
libzopfli=1.0.3=he1b5a44_0
llvmlite=0.33.0=py38hc6ec683_1
lz4-c=1.9.2=he6710b0_1
markdown=3.2.2=py_0
matplotlib=3.3.1=1
matplotlib-base=3.3.1=py38h817c723_0
mccabe=0.6.1=py38_1
mkl=2020.1=217
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
mpl_sample_data=3.3.1=py_0
ncurses=6.2=he6710b0_1
netifaces=0.10.9=py38h1e0a361_1002
networkx=2.5=py_0
ninja=1.10.0=py38hfd86e86_0
numba=0.50.1=py38h0573a6f_1
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
nvidia-ml=7.352.0=py_0
olefile=0.46=py_0
openjpeg=2.3.1=h981e76c_3
openssl=1.1.1h=h516909a_0
owslib=0.20.0=py_0
pcre=8.44=he6710b0_0
pillow=7.2.0=py38hb39fc2d_0
pip=20.2.1=py38_0
pixman=0.38.0=h516909a_1003
poppler=0.89.0=h4190859_1
poppler-data=0.4.9=1
postgresql=12.3=h8573dbc_0
proj=7.1.0=h966b41f_1
protobuf=3.12.4=py38h950e882_0
psutil=5.7.0=py38h7b6447c_0
pycparser=2.20=py_2
pyepsg=0.4.0=py_0
pylint=2.5.3=py38_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyproj=2.6.1.post1=py38h8e47818_1
pyrsistent=0.16.0=py38h7b6447c_0
pyshp=2.1.0=py_0
pysocks=1.7.1=py38_0
python=3.8.5=hcff3b4d_0
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=0.8.5=py_0
pytz=2020.1=pyh9f0ad1d_0
pywavelets=1.1.1=py38h8790de6_1
pyyaml=5.3.1=py38h1e0a361_0
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
scikit-image=0.17.2=py38hcb8c335_1
scikit-learn=0.23.1=py38h423224d_0
scipy=1.5.0=py38h0b6359f_0
setuptools=49.2.1=py38_0
shapely=1.7.1=py38hc7361b7_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
snappy=1.1.8=he1b5a44_3
snuggs=1.4.7=py_0
sqlite=3.32.3=h62c20be_0
tbb=2020.1=hc9558a2_0
tensorboard=1.15.0=py38_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.8.13=py_0
tiledb=2.0.8=h3effe38_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
toolz=0.10.0=py_0
torchvision=0.7.0=py38_cu101
tornado=6.0.4=py38h7b6447c_1
tqdm=4.48.2=pyh9f0ad1d_0
tzcode=2020a=h516909a_0
urllib3=1.25.10=py_0
websocket-client=0.57.0=py38_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py38_0
wrapt=1.11.2=py38h7b6447c_0
wurlitzer=2.0.1=py38_0
xerces-c=3.2.3=hfe33f54_1
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h516909a_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
